Debugger: Fixed disassembler incorrectly calculating jump destinations and caching the wrong data as a result

+++ Core/DisassemblyInfo.h
@@ -34,7 +34,7 @@ public:
 	
 	void GetEffectiveAddressString(string &out, State& cpuState, MemoryManager* memoryManager, LabelManager* labelManager);
 	int32_t GetMemoryValue(State& cpuState, MemoryManager* memoryManager);
-	int32_t GetJumpDestination(uint16_t pc, MemoryManager* memoryManager);
+	uint16_t GetJumpDestination(uint16_t pc, MemoryManager* memoryManager);
 	uint16_t GetIndirectJumpDestination(MemoryManager* memoryManager);
 	void ToString(string &out, uint32_t memoryAddr, MemoryManager* memoryManager, LabelManager* labelManager, bool extendZeroPage);
 	void GetByteCode(string &out);
